[REFACTOR] Consolidate ownership data and validation
