+++ src/tools/definition.ts
@@ -4,11 +4,9 @@ import { AssetSourcePath, DefaultId } from './context';
 import { LoadAssetsGraphics } from './graphics';
 import { GraphicsDatabase } from './graphicsDatabase';
 import { join } from 'path';
-import { Contributors, CurrentCommitter, GitDataAvailable } from './git';
-import { IS_PRODUCTION_BUILD } from '../constants';
-import * as fs from 'fs';
 import { pick } from 'lodash';
 import { LoadAssetColorization } from './load_helpers/color';
+import { ValidateOwnershipData } from './licensing';
 
 const DEFINITION_FALLTHROUGH_PROPERTIES = [
 	// Properties
@@ -63,47 +61,8 @@ export function GlobalDefineAsset(def: IntermediatePersonalAssetDefinition): voi
 		ValidateAssetDefinitionPoseLimits(logger, 'poseLimits', def.poseLimits);
 	}
 
-	//#region Validate ownership data
-
-	// Validate responsible contributor
-	const contributor = def.ownership.responsibleContributor.toLowerCase();
-	if (GitDataAvailable &&
-		!Contributors.has(contributor) &&
-		(!CurrentCommitter || CurrentCommitter.toLowerCase() !== contributor)
-	) {
-		if (IS_PRODUCTION_BUILD || !CurrentCommitter) {
-			logger.warning('The responsible contributor was not found in the Git history.');
-		} else {
-			logger.warning(
-				'The responsible contributor was not found in the Git history.\n' +
-				`If you commit with current settings, this is your commit signature: '${CurrentCommitter}'`,
-			);
-		}
-	}
-
-	// Validate presence of licensing data
-	if (def.graphics !== undefined && def.ownership.licensing.length === 0) {
-		logger.warning('Asset has graphics, but no licensing info');
-	}
-
-	for (const license of def.ownership.licensing) {
-		// Validate that custom license exists and is a file
-		if (license.license.startsWith('./')) {
-			const path = join(AssetSourcePath, license.license);
-			if (!fs.existsSync(path) || !fs.statSync(path).isFile()) {
-				logger.warning(`Custom license '${license.license}' doesn't exist or is not a file.`);
-			}
-		}
-	}
-	// Check that CHANGE_ME was replaced
-	if (def.ownership.licensing
-		.flatMap((l) => [l.part, l.copyrightHolder, l.editedBy])
-		.includes('CHANGE_ME')
-	) {
-		logger.warning(`Licensing data includes fields with 'CHANGE_ME' template data that need to be changed.`);
-	}
-
-	//#endregion
+	// Validate ownership data
+	ValidateOwnershipData(def.ownership, logger, def.graphics != null);
 
 	if (!definitionValid) {
 		logger.error('Invalid asset definition, asset skipped');
