+++ src/types.d.ts
@@ -36,26 +36,29 @@ type IntermediateAssetColorization
 	= (Omit<AssetColorizationBetterMinAlpha, 'group'> & { group?: never; })
 	| (Omit<AssetColorizationBetterMinAlpha, 'default'> & { group: AssetRepoExtraArgs['colorGroups']; default?: never; });
 
+/** Information about asset ownership and author's requests about the asset */
+type AssetOwnershipData = {
+	/** Same as author of git commits present in PR, has responsibility for this asset */
+	responsibleContributor: string;
+	/** Who is shown in credits for this asset and at the same time people to ask when Asset should be changed */
+	credits: string[];
+	/** Defines author's preferences about how their asset can be modified */
+	modificationPolicy: PandoraAssetModificationPolicy;
+	/** Defines author's preferences about how their asset can be reused for other assets */
+	reusePolicy: PandoraAssetReusePolicy;
+	/**
+	 * Legal info about the images
+	 * If there are multiple sources used, specify this multiple times.
+	 * If author gave you express permission to use images, but wishes to remain Anonymous, write "Anonymous" into relevant fields.
+	 */
+	licensing: LicensingInfo[];
+};
+
 interface IntermediatePersonalAssetDefinition extends Pick<import('pandora-common').PersonalAssetDefinition<AssetRepoExtraArgs>, import('./tools/definition').AssetDefinitionFallthroughProperties> {
 	id?: string;
 	graphics?: string;
 	/** Info about who owns the asset(s) */
-	ownership: {
-		/** Same as author of git commits present in PR, has responsibility for this asset */
-		responsibleContributor: string;
-		/** Who is shown in credits for this asset and at the same time people to ask when Asset should be changed */
-		credits: string[];
-		/** Defines author's preferences about how their asset can be modified */
-		modificationPolicy: PandoraAssetModificationPolicy;
-		/** Defines author's preferences about how their asset can be reused for other assets */
-		reusePolicy: PandoraAssetReusePolicy;
-		/**
-		 * Legal info about the images
-		 * If there are multiple sources used, specify this multiple times.
-		 * If author gave you express permission to use images, but wishes to remain Anonymous, write "Anonymous" into relevant fields.
-		 */
-		licensing: LicensingInfo[];
-	};
+	ownership: AssetOwnershipData;
 	colorization?: Record<string, IntermediateAssetColorization>;
 }
 
@@ -72,43 +75,13 @@ interface IntermediateRoomDeviceDefinition extends Pick<import('pandora-common')
 	id?: string;
 	slots: Record<string, IntermediateRoomDeviceSlotDefinition>;
 	/** Info about who owns the asset(s) */
-	ownership: {
-		/** Same as author of git commits present in PR, has responsibility for this asset */
-		responsibleContributor: string;
-		/** Who is shown in credits for this asset and at the same time people to ask when Asset should be changed */
-		credits: string[];
-		/** Defines author's preferences about how their asset can be modified */
-		modificationPolicy: PandoraAssetModificationPolicy;
-		/** Defines author's preferences about how their asset can be reused for other assets */
-		reusePolicy: PandoraAssetReusePolicy;
-		/**
-		 * Legal info about the images
-		 * If there are multiple sources used, specify this multiple times.
-		 * If author gave you express permission to use images, but wishes to remain Anonymous, write "Anonymous" into relevant fields.
-		 */
-		licensing: LicensingInfo[];
-	};
+	ownership: AssetOwnershipData;
 }
 
 interface IntermediateLockAssetDefinition extends Pick<import('pandora-common').LockAssetDefinition, import('./tools/definitionLock').LockAssetDefinitionFallthroughProperties> {
 	id?: string;
 	/** Info about who owns the asset(s) */
-	ownership: {
-		/** Same as author of git commits present in PR, has responsibility for this asset */
-		responsibleContributor: string;
-		/** Who is shown in credits for this asset and at the same time people to ask when Asset should be changed */
-		credits: string[];
-		/** Defines author's preferences about how their asset can be modified */
-		modificationPolicy: PandoraAssetModificationPolicy;
-		/** Defines author's preferences about how their asset can be reused for other assets */
-		reusePolicy: PandoraAssetReusePolicy;
-		/**
-		 * Legal info about the images
-		 * If there are multiple sources used, specify this multiple times.
-		 * If author gave you express permission to use images, but wishes to remain Anonymous, write "Anonymous" into relevant fields.
-		 */
-		licensing: LicensingInfo[];
-	};
+	ownership: AssetOwnershipData;
 }
 
 interface IntermediateRoomBackgroundDefinition extends Pick<import('pandora-common').IChatroomBackgroundInfo,
